requirements.txt
streamlit
duckdb
pandas
plotly
pydeck
google-generativeai
langdetect
requests
